NODE-782: Move classes CLI and DockerCLI to library
